Move broker, publish and sbscribe slices to slices folder
